pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "folio-trainer"
version = "0.1.0"
description = "Point-in-time daily predictive asset allocation pipeline"
readme = "docs/allocation_model_spec.md"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.0",
    "pyarrow>=14",
    "pydantic>=2.5",
    "pyyaml>=6",
    "click>=8",
    "exchange-calendars>=4",
    "numpy>=1.26",
    "scipy>=1.12",
    "lightgbm>=4",
    "optuna>=3",
    "requests>=2.31",
    "fredapi>=0.5",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
nn = ["torch>=2"]
dev = [
    "pytest>=8",
    "pytest-cov>=4",
    "ruff>=0.4",
]

[project.scripts]
folio-trainer = "folio_trainer.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/folio_trainer"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
